added player related helpers

+++ src/storm/storage/prism/Player.cpp
@@ -2,7 +2,7 @@
 
 namespace storm {
     namespace prism {
-        Player::Player(std::string const& playerName, std::map<std::string, uint_fast64_t> const& controlledModules, std::map<std::string, uint_fast64_t> const& controlledCommands, std::string const& filename, uint_fast64_t lineNumber) : LocatedInformation(filename, lineNumber), playerName(playerName), controlledModules(controlledModules), controlledCommands(controlledCommands) {
+        Player::Player(std::string const& playerName, std::map<std::string, uint_fast32_t> const& controlledModules, std::map<std::string, uint_fast32_t> const& controlledCommands, std::string const& filename, uint_fast32_t lineNumber) : LocatedInformation(filename, lineNumber), playerName(playerName), controlledModules(controlledModules), controlledCommands(controlledCommands) {
             // Nothing to do here.
         }
 
@@ -10,11 +10,11 @@ namespace storm {
             return this->playerName;
         }
 
-        std::map<std::string, uint_fast64_t> const& Player::getModules() const {
+        std::map<std::string, uint_fast32_t> const& Player::getModules() const {
             return this->controlledModules;
         }
 
-        std::map<std::string, uint_fast64_t> const& Player::getCommands() const {
+        std::map<std::string, uint_fast32_t> const& Player::getCommands() const {
             return this->controlledCommands;
         }
